fix3

+++ src/ru/geekbrains/lesson1/CatRun.java
@@ -2,7 +2,7 @@ package ru.geekbrains.lesson1;
 
 public class CatRun implements RunBehavior {
     @Override
-    public int run(int length) {
-        return length / 5;
+    public int run(double length) {
+        return (int)length / 5;
     }
 }
