+++ src/ru/geekbrains/lesson1/HumanRun.java
@@ -2,7 +2,7 @@ package ru.geekbrains.lesson1;
 
 public class HumanRun implements RunBehavior {
     @Override
-    public int run(int length) {
-        return length;
+    public int run(double length) {
+        return (int)length;
     }
 }
