+++ src/ru/geekbrains/lesson1/RobotRun.java
@@ -2,7 +2,7 @@ package ru.geekbrains.lesson1;
 
 public class RobotRun implements RunBehavior {
     @Override
-    public int run(int length) {
-        return length * 2;
+    public int run(double length) {
+        return (int)length * 2;
     }
 }
