+++ src/ru/geekbrains/lesson1/RunBehavior.java
@@ -1,5 +1,5 @@
 package ru.geekbrains.lesson1;
 
 public interface RunBehavior {
-    int run(int height);
+    int run(double height);
 }
